requirements.txt
anyio==3.7.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.104.1
google==3.0.0
google-auth==2.25.2
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlalchemy2-stubs==0.0.2a37
sqlmodel==0.0.11
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uuid==1.30
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
